s3-tar-converter1/requirements.txt
python-dotenv==0.10.3
cryptography==2.8
